+++ src/pkg/services/m365/api/shared.go
